✅ Merge back `frequency` specs in one file
